+++ model_navigator/framework_api/commands/convert/tf.py
@@ -115,7 +115,7 @@ class ConvertSavedModel2TFTRT(ConvertBase):
             self.status = Status.SKIPPED
             return
 
-        with ExecutionContext(converted_model_path.parent / "reproduce.py") as context:
+        with ExecutionContext(converted_model_path.parent / "reproduce_conversion.py") as context:
             kwargs = {
                 "exported_model_path": exported_model_path.as_posix(),
                 "converted_model_path": converted_model_path.as_posix(),