+++ model_navigator/framework_api/contrib/huggingface/torch/export.py
@@ -130,7 +130,7 @@ def export(
     batch_dim: Optional[int] = 0,
     padding: Union[bool, str] = True,
     max_sequence_len: Optional[int] = None,
-    onnx_runtimes: Optional[Union[Union[str, RuntimeProvider], Tuple[Union[str, RuntimeProvider], ...]]] = None,
+    runtimes: Optional[Union[Union[str, RuntimeProvider], Tuple[Union[str, RuntimeProvider], ...]]] = None,
     run_profiling: bool = True,
     profiler_config: Optional[ProfilerConfig] = None,
 ) -> PackageDescriptor:
@@ -204,16 +204,16 @@ def export(
         forward_kw_names = tuple(sample.keys())
     else:
         forward_kw_names = None
-    onnx_runtimes = onnx_runtimes or format2runtimes(Format.ONNX)
+    runtimes = runtimes or format2runtimes(Format.ONNX)
 
     if profiler_config is None:
         profiler_config = ProfilerConfig()
 
-    target_formats, jit_options, target_precisions, onnx_runtimes, precision_mode = (
+    target_formats, jit_options, target_precisions, runtimes, precision_mode = (
         parse_enum(target_formats, Format),
         parse_enum(jit_options, JitType),
         parse_enum(target_precisions, TensorRTPrecision),
-        parse_enum(onnx_runtimes, RuntimeProvider),
+        parse_enum(runtimes, RuntimeProvider),
         *parse_enum(precision_mode, TensorRTPrecisionMode),
     )
 
@@ -241,7 +241,7 @@ def export(
         target_device=target_device,
         disable_git_info=disable_git_info,
         batch_dim=batch_dim,
-        onnx_runtimes=onnx_runtimes,
+        runtimes=runtimes,
         profiler_config=profiler_config,
     )