+++ model_navigator/framework_api/onnx.py
@@ -57,7 +57,7 @@ def export(
     rtol: Optional[float] = None,
     disable_git_info: bool = False,
     batch_dim: Optional[int] = 0,
-    onnx_runtimes: Optional[Union[Union[str, RuntimeProvider], Tuple[Union[str, RuntimeProvider], ...]]] = None,
+    runtimes: Optional[Union[Union[str, RuntimeProvider], Tuple[Union[str, RuntimeProvider], ...]]] = None,
     run_profiling: bool = True,
     profiler_config: Optional[ProfilerConfig] = None,
 ) -> PackageDescriptor:
@@ -82,16 +82,16 @@ def export(
     if sample_count is None:
         sample_count = 100
 
-    if onnx_runtimes is None:
-        onnx_runtimes = format2runtimes(Format.ONNX)
+    if runtimes is None:
+        runtimes = format2runtimes(Format.ONNX)
 
     if profiler_config is None:
         profiler_config = ProfilerConfig()
 
-    target_formats, target_precisions, onnx_runtimes = (
+    target_formats, target_precisions, runtimes = (
         parse_enum(target_formats, Format),
         parse_enum(target_precisions, TensorRTPrecision),
-        parse_enum(onnx_runtimes, RuntimeProvider),
+        parse_enum(runtimes, RuntimeProvider),
     )
     config = Config(
         Framework.ONNX,
@@ -110,7 +110,7 @@ def export(
         rtol=rtol,
         disable_git_info=disable_git_info,
         batch_dim=batch_dim,
-        onnx_runtimes=onnx_runtimes,
+        runtimes=runtimes,
         profiler_config=profiler_config,
     )