+++ model_navigator/framework_api/runners/tf.py
@@ -22,10 +22,11 @@ import tensorflow as tf  # pytype: disable=import-error
 from polygraphy.backend.base import BaseRunner
 from polygraphy.common import TensorMetadata
 
+from model_navigator.framework_api.runners.base import INavigatorRunner
 from model_navigator.framework_api.utils import Framework, validate_sample_output
 
 
-class TFBaseRunner(BaseRunner):
+class TFBaseRunner(INavigatorRunner, BaseRunner):
     """
     Runs inference using TensorFlow2.
     """